feature LikedCommunity

+++ hobbyhobby-community/hobbyhobby-domain/src/main/java/org/v1/repository/UserCommunityRepository.java
@@ -3,6 +3,6 @@ package org.v1.repository;
 import java.util.List;
 
 public interface UserCommunityRepository {
-    void appendUserToCommunity(Long userId, Integer communityId);
-    void removeUserFromCommunity(Long userId, Integer communityId);
+    void appendUserToCommunity(Long userId, Long communityId);
+    void removeUserFromCommunity(Long userId, Long communityId);
 }
